Move common code to base.py
